refactor(sr_od): app/{div_uni,treasures_lightward} 迁移到 application/
完成 sr_od 业务代码向 application/ 架构的统一：
- app/div_uni → application/div_uni（5 文件）
- app/treasures_lightward → application/treasures_lightward（18 文件）
- 删除 app/sr_application_launcher.py（转发 stub，application/ 已有真实实现）+ 空 app/__init__.py
- 导入路径 sr_od.app.{div_uni,treasures_lightward} → sr_od.application.*
- 清理空 app/ 目录
