--- src/sr_od/app/sr_application_launcher.py
@@ -1,6 +1,0 @@
-from sr_od.application.sr_application_launcher import SrApplicationLauncher
-
-
-if __name__ == '__main__':
-    launcher = SrApplicationLauncher()
-    launcher.run()
