+++ src/sr_od/application/div_uni/operations/ornamenet_extraction.py
@@ -8,8 +8,8 @@ from one_dragon.base.operation.operation_round_result import OperationRoundResul
 from one_dragon.utils import str_utils, cv2_utils
 from one_dragon.utils.i18_utils import gt
 from one_dragon.utils.log_utils import log
-from sr_od.app.div_uni.operations.choose_oe_file import ChooseOeFile
-from sr_od.app.div_uni.operations.choose_oe_support import ChooseOeSupport
+from sr_od.application.div_uni.operations.choose_oe_file import ChooseOeFile
+from sr_od.application.div_uni.operations.choose_oe_support import ChooseOeSupport
 from sr_od.application.sim_universe.operations.move_v1.sim_uni_move_to_enemy_by_mm import SimUniMoveToEnemyByMiniMap
 from sr_od.challenge_mission.choose_challenge_times import ChooseChallengeTimes
 from sr_od.context.sr_context import SrContext
