+++ src/sr_od/application/treasures_lightward/op/check_star.py
